VMs was moved to CV cs files
