+++ Src/English.Maui/Controls/ObjectCV.xaml
@@ -1,12 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:controls="clr-namespace:English.Maui.Controls"
              x:Class="English.Maui.Controls.ObjectCV">
-
-    <ContentView.BindingContext>
-        <controls:ObjectVM />
-    </ContentView.BindingContext>
     <VerticalStackLayout >
         <StackLayout Orientation="Horizontal">
             <Picker  Margin="10,1"
