+++ Src/English.Maui/Controls/VerbCV.xaml
@@ -1,13 +1,9 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:controls="clr-namespace:English.Maui.Controls"
              x:Class="English.Maui.Controls.VerbCV">
 
 
-    <ContentView.BindingContext>
-        <controls:VerbVM />
-    </ContentView.BindingContext>
 
     <StackLayout Orientation="Horizontal" >
         <StackLayout WidthRequest="400">
